+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/ChunkyDatagen.java
@@ -7,31 +7,37 @@ import com.mrbysco.chunkymcchunkface.datagen.data.ChunkyBlockTagProvider;
 import com.mrbysco.chunkymcchunkface.datagen.data.ChunkyItemTagProvider;
 import com.mrbysco.chunkymcchunkface.datagen.data.ChunkyLootProvider;
 import com.mrbysco.chunkymcchunkface.datagen.data.ChunkyRecipeProvider;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.data.DataGenerator;
-import net.minecraft.data.tags.BlockTagsProvider;
+import net.minecraft.data.PackOutput;
+import net.minecraftforge.common.data.BlockTagsProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.data.event.GatherDataEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
+
+import java.util.concurrent.CompletableFuture;
 
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
 public class ChunkyDatagen {
 	@SubscribeEvent
 	public static void gatherData(GatherDataEvent event) {
 		DataGenerator generator = event.getGenerator();
+		PackOutput packOutput = generator.getPackOutput();
+		CompletableFuture<HolderLookup.Provider> lookupProvider = event.getLookupProvider();
 		ExistingFileHelper helper = event.getExistingFileHelper();
 
 		if (event.includeServer()) {
-			generator.addProvider(true, new ChunkyRecipeProvider(generator));
-			generator.addProvider(true, new ChunkyLootProvider(generator));
+			generator.addProvider(true, new ChunkyRecipeProvider(packOutput));
+			generator.addProvider(true, new ChunkyLootProvider(packOutput));
 			BlockTagsProvider provider;
-			generator.addProvider(true, provider = new ChunkyBlockTagProvider(generator, helper));
-			generator.addProvider(true, new ChunkyItemTagProvider(generator, provider, helper));
+			generator.addProvider(true, provider = new ChunkyBlockTagProvider(packOutput, lookupProvider, helper));
+			generator.addProvider(true, new ChunkyItemTagProvider(packOutput, lookupProvider, provider, helper));
 		}
 		if (event.includeClient()) {
-			generator.addProvider(true, new ChunkyLanguageProvider(generator));
-			generator.addProvider(true, new ChunkyStateProvider(generator, helper));
-			generator.addProvider(true, new ChunkyItemModelProvider(generator, helper));
+			generator.addProvider(true, new ChunkyLanguageProvider(packOutput));
+			generator.addProvider(true, new ChunkyStateProvider(packOutput, helper));
+			generator.addProvider(true, new ChunkyItemModelProvider(packOutput, helper));
 		}
 	}
 }