+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/assets/ChunkyItemModelProvider.java
@@ -3,13 +3,14 @@ package com.mrbysco.chunkymcchunkface.datagen.assets;
 import com.mrbysco.chunkymcchunkface.ChunkyMcChunkFace;
 import com.mrbysco.chunkymcchunkface.registry.ChunkyRegistry;
 import net.minecraft.data.DataGenerator;
+import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.client.model.generators.ItemModelProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
 
 public class ChunkyItemModelProvider extends ItemModelProvider {
-	public ChunkyItemModelProvider(DataGenerator gen, ExistingFileHelper helper) {
-		super(gen, ChunkyMcChunkFace.MOD_ID, helper);
+	public ChunkyItemModelProvider(PackOutput packOutput, ExistingFileHelper helper) {
+		super(packOutput, ChunkyMcChunkFace.MOD_ID, helper);
 	}
 
 	@Override