+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/assets/ChunkyStateProvider.java
@@ -3,13 +3,14 @@ package com.mrbysco.chunkymcchunkface.datagen.assets;
 import com.mrbysco.chunkymcchunkface.ChunkyMcChunkFace;
 import com.mrbysco.chunkymcchunkface.registry.ChunkyRegistry;
 import net.minecraft.data.DataGenerator;
+import net.minecraft.data.PackOutput;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraftforge.client.model.generators.BlockStateProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
 
 public class ChunkyStateProvider extends BlockStateProvider {
-	public ChunkyStateProvider(DataGenerator gen, ExistingFileHelper helper) {
-		super(gen, ChunkyMcChunkFace.MOD_ID, helper);
+	public ChunkyStateProvider(PackOutput packOutput, ExistingFileHelper helper) {
+		super(packOutput, ChunkyMcChunkFace.MOD_ID, helper);
 	}
 
 	@Override