+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/data/ChunkyBlockTagProvider.java
@@ -2,19 +2,22 @@ package com.mrbysco.chunkymcchunkface.datagen.data;
 
 import com.mrbysco.chunkymcchunkface.ChunkyMcChunkFace;
 import com.mrbysco.chunkymcchunkface.registry.ChunkyTags;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.tags.BlockTagsProvider;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
 import net.minecraft.tags.BlockTags;
+import net.minecraftforge.common.data.BlockTagsProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
-import org.jetbrains.annotations.Nullable;
+
+import javax.annotation.Nullable;
+import java.util.concurrent.CompletableFuture;
 
 public class ChunkyBlockTagProvider extends BlockTagsProvider {
-	public ChunkyBlockTagProvider(DataGenerator generator, @Nullable ExistingFileHelper fileHelper) {
-		super(generator, ChunkyMcChunkFace.MOD_ID, fileHelper);
+	public ChunkyBlockTagProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, @Nullable ExistingFileHelper existingFileHelper) {
+		super(output, lookupProvider, ChunkyMcChunkFace.MOD_ID, existingFileHelper);
 	}
 
 	@Override
-	protected void addTags() {
+	protected void addTags(HolderLookup.Provider provider) {
 		this.tag(ChunkyTags.UPGRADE_BLOCKS).addTags(BlockTags.BEACON_BASE_BLOCKS);
 	}
 }