+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/data/ChunkyItemTagProvider.java
@@ -1,18 +1,23 @@
 package com.mrbysco.chunkymcchunkface.datagen.data;
 
 import com.mrbysco.chunkymcchunkface.ChunkyMcChunkFace;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.tags.BlockTagsProvider;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
+import net.minecraft.data.tags.TagsProvider;
+import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.data.ExistingFileHelper;
 
+import java.util.concurrent.CompletableFuture;
+
 public class ChunkyItemTagProvider extends ItemTagsProvider {
-	public ChunkyItemTagProvider(DataGenerator dataGenerator, BlockTagsProvider blockTagsProvider, ExistingFileHelper fileHelper) {
-		super(dataGenerator, blockTagsProvider, ChunkyMcChunkFace.MOD_ID, fileHelper);
+	public ChunkyItemTagProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider,
+								 TagsProvider<Block> blockTagProvider, ExistingFileHelper existingFileHelper) {
+		super(output, lookupProvider, blockTagProvider, ChunkyMcChunkFace.MOD_ID, existingFileHelper);
 	}
 
 	@Override
-	protected void addTags() {
+	protected void addTags(HolderLookup.Provider provider) {
 
 	}
 }