+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/data/ChunkyRecipeProvider.java
@@ -1,8 +1,9 @@
 package com.mrbysco.chunkymcchunkface.datagen.data;
 
 import com.mrbysco.chunkymcchunkface.registry.ChunkyRegistry;
-import net.minecraft.data.DataGenerator;
+import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.FinishedRecipe;
+import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.data.recipes.RecipeProvider;
 import net.minecraft.data.recipes.ShapedRecipeBuilder;
 import net.minecraft.world.item.Items;
@@ -11,13 +12,13 @@ import net.minecraftforge.common.Tags;
 import java.util.function.Consumer;
 
 public class ChunkyRecipeProvider extends RecipeProvider {
-	public ChunkyRecipeProvider(DataGenerator gen) {
-		super(gen);
+	public ChunkyRecipeProvider(PackOutput packOutput) {
+		super(packOutput);
 	}
 
 	@Override
-	protected void buildCraftingRecipes(Consumer<FinishedRecipe> recipeConsumer) {
-		ShapedRecipeBuilder.shaped(ChunkyRegistry.CHUNK_LOADER.get())
+	protected void buildRecipes(Consumer<FinishedRecipe> recipeConsumer) {
+		ShapedRecipeBuilder.shaped(RecipeCategory.REDSTONE, ChunkyRegistry.CHUNK_LOADER.get())
 				.define('G', Items.GOLD_BLOCK)
 				.define('E', Tags.Items.NETHER_STARS)
 				.define('O', Tags.Items.OBSIDIAN)