+++ src/main/java/com/mrbysco/chunkymcchunkface/registry/ChunkyRegistry.java
@@ -25,6 +25,6 @@ public class ChunkyRegistry {
 	public static final RegistryObject<BlockEntityType<ChunkLoaderBlockEntity>> CHUNK_LOADER_ENTITY = BLOCK_ENTITIES.register("chunk_loader", () ->
 			BlockEntityType.Builder.of(ChunkLoaderBlockEntity::new, CHUNK_LOADER.get()).build(null));
 	public static final RegistryObject<Item> CHUNK_LOADER_ITEM = ITEMS.register("chunk_loader", () -> new BlockItem(CHUNK_LOADER.get(),
-			new Item.Properties().tab(CreativeModeTab.TAB_REDSTONE)));
+			new Item.Properties()));
 
 }