+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/assets/ChunkyItemModelProvider.java
@@ -2,7 +2,6 @@ package com.mrbysco.chunkymcchunkface.datagen.assets;
 
 import com.mrbysco.chunkymcchunkface.ChunkyMcChunkFace;
 import com.mrbysco.chunkymcchunkface.registry.ChunkyRegistry;
-import net.minecraft.data.DataGenerator;
 import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.client.model.generators.ItemModelProvider;