+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/assets/ChunkyStateProvider.java
@@ -2,7 +2,6 @@ package com.mrbysco.chunkymcchunkface.datagen.assets;
 
 import com.mrbysco.chunkymcchunkface.ChunkyMcChunkFace;
 import com.mrbysco.chunkymcchunkface.registry.ChunkyRegistry;
-import net.minecraft.data.DataGenerator;
 import net.minecraft.data.PackOutput;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraftforge.client.model.generators.BlockStateProvider;