+++ src/main/java/com/mrbysco/chunkymcchunkface/datagen/data/ChunkyLootProvider.java
@@ -8,7 +8,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.storage.loot.LootTable;
-import net.minecraft.world.level.storage.loot.LootTables;
 import net.minecraft.world.level.storage.loot.ValidationContext;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
 import net.minecraftforge.registries.RegistryObject;
@@ -42,6 +41,6 @@ public class ChunkyLootProvider extends LootTableProvider {
 
 	@Override
 	protected void validate(Map<ResourceLocation, LootTable> map, ValidationContext validationContext) {
-		map.forEach((name, table) -> LootTables.validate(validationContext, name, table));
+		map.forEach((name, table) -> table.validate(validationContext));
 	}
 }