+++ src/main/java/com/mrbysco/chunkymcchunkface/handler/PlayerHandler.java
@@ -11,7 +11,7 @@ public class PlayerHandler {
 	@SubscribeEvent
 	public void onPlayerLogin(PlayerEvent.PlayerLoggedInEvent event) {
 		Player player = event.getEntity();
-		Level level = player.level;
+		Level level = player.level();
 		if (!level.isClientSide) {
 			ChunkData data = ChunkData.get(level);
 			data.removePlayer(player.getUUID());
@@ -21,7 +21,7 @@ public class PlayerHandler {
 	@SubscribeEvent
 	public void onPlayerLogout(PlayerEvent.PlayerLoggedOutEvent event) {
 		Player player = event.getEntity();
-		Level level = player.level;
+		Level level = player.level();
 		if (!level.isClientSide) {
 			ChunkData data = ChunkData.get(level);
 			data.addPlayer(player.getUUID(), level.getGameTime());
@@ -31,7 +31,7 @@ public class PlayerHandler {
 	@SubscribeEvent
 	public void onPlayerTick(TickEvent.PlayerTickEvent event) {
 		Player player = event.player;
-		Level level = player.level;
+		Level level = player.level();
 		if (!level.isClientSide && level.getGameTime() % 20L == 0L) {
 			ChunkData data = ChunkData.get(level);
 			if (level.getGameTime() > data.getLastSeen(player.getUUID())) {