+++ src/main/java/com/mrbysco/chunkymcchunkface/registry/ChunkyRegistry.java
@@ -4,12 +4,11 @@ import com.mrbysco.chunkymcchunkface.ChunkyMcChunkFace;
 import com.mrbysco.chunkymcchunkface.blocks.ChunkLoaderBlock;
 import com.mrbysco.chunkymcchunkface.blocks.entity.ChunkLoaderBlockEntity;
 import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.entity.BlockEntityType;
-import net.minecraft.world.level.material.Material;
+import net.minecraft.world.level.material.MapColor;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
@@ -20,7 +19,7 @@ public class ChunkyRegistry {
 	public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, ChunkyMcChunkFace.MOD_ID);
 
 	public static final RegistryObject<Block> CHUNK_LOADER = BLOCKS.register("chunk_loader", () ->
-			new ChunkLoaderBlock(Block.Properties.of(Material.DECORATION).strength(0.8F).sound(SoundType.METAL).noOcclusion()));
+			new ChunkLoaderBlock(Block.Properties.of().mapColor(MapColor.GOLD).strength(0.8F).sound(SoundType.METAL).noOcclusion()));
 
 	public static final RegistryObject<BlockEntityType<ChunkLoaderBlockEntity>> CHUNK_LOADER_ENTITY = BLOCK_ENTITIES.register("chunk_loader", () ->
 			BlockEntityType.Builder.of(ChunkLoaderBlockEntity::new, CHUNK_LOADER.get()).build(null));