+++ changelog.md
@@ -1,1 +1,1 @@
-* Initial 1.19.4 build
+* Initial update to 1.20